+++ src/TaskEither.ts
@@ -232,28 +232,27 @@ export const toUnion: <E, A>(fa: TaskEither<E, A>) => Task<E | A> = /*#__PURE__*
  * @category interop
  * @since 3.0.0
  */
-export const fromNullable: <E>(e: E) => <A>(a: A) => TaskEither<E, NonNullable<A>> = /*#__PURE__*/ ET.fromNullable(
-  T.Pointed
-)
+export const fromNullableOrElse: <E>(onNullable: Lazy<E>) => <A>(a: A) => TaskEither<E, NonNullable<A>> =
+  /*#__PURE__*/ ET.fromNullableOrElse(T.Pointed)
 
 /**
  * @category interop
  * @since 3.0.0
  */
-export const fromNullableK: <E>(
-  e: E
+export const fromNullableKOrElse: <E>(
+  onNullable: Lazy<E>
 ) => <A extends ReadonlyArray<unknown>, B>(
   f: (...a: A) => B | null | undefined
-) => (...a: A) => TaskEither<E, NonNullable<B>> = /*#__PURE__*/ ET.fromNullableK(T.Pointed)
+) => (...a: A) => TaskEither<E, NonNullable<B>> = /*#__PURE__*/ ET.fromNullableKOrElse(T.Pointed)
 
 /**
  * @category interop
  * @since 3.0.0
  */
-export const chainNullableK: <E>(
-  e: E
+export const chainNullableKOrElse: <E>(
+  onNullable: Lazy<E>
 ) => <A, B>(f: (a: A) => B | null | undefined) => (ma: TaskEither<E, A>) => TaskEither<E, NonNullable<B>> =
-  /*#__PURE__*/ ET.chainNullableK(T.Monad)
+  /*#__PURE__*/ ET.chainNullableKOrElse(T.Monad)
 
 // -------------------------------------------------------------------------------------
 // combinators